move data layer code under dataLayer
