+++ src/dataLayer/databaseHistoryManager.js
