+++ src/dataLayer/databaseQueryManager.js
