+++ src/dataLayer/databaseUpdateManager.js
